fix: roting and link constat
