+++ src/app/(Page)/pokemon/search/page.tsx
