+++ src/app/constants/index.ts
@@ -1,1005 +1,13 @@
 /**
- * ポケモンの日本語名とそのid
+ * アプリ内のリンクリスト
  */
-export const POKEMON_JAPANESE_NAMES = [
-  { id: 1, name: "フシギダネ" },
-  { id: 2, name: "フシギソウ" },
-  { id: 3, name: "フシギバナ" },
-  { id: 4, name: "ヒトカゲ" },
-  { id: 5, name: "リザード" },
-  { id: 6, name: "リザードン" },
-  { id: 7, name: "ゼニガメ" },
-  { id: 8, name: "カメール" },
-  { id: 9, name: "カメックス" },
-  { id: 10, name: "キャタピー" },
-  { id: 11, name: "トランセル" },
-  { id: 12, name: "バタフリー" },
-  { id: 13, name: "ビードル" },
-  { id: 14, name: "コクーン" },
-  { id: 15, name: "スピアー" },
-  { id: 16, name: "ポッポ" },
-  { id: 17, name: "ピジョン" },
-  { id: 18, name: "ピジョット" },
-  { id: 19, name: "コラッタ" },
-  { id: 20, name: "ラッタ" },
-  { id: 21, name: "オニスズメ" },
-  { id: 22, name: "オニドリル" },
-  { id: 23, name: "アーボ" },
-  { id: 24, name: "アーボック" },
-  { id: 25, name: "ピカチュウ" },
-  { id: 26, name: "ライチュウ" },
-  { id: 27, name: "サンド" },
-  { id: 28, name: "サンドパン" },
-  { id: 29, name: "ニドラン♀" },
-  { id: 30, name: "ニドリーナ" },
-  { id: 31, name: "ニドクイン" },
-  { id: 32, name: "ニドラン♂" },
-  { id: 33, name: "ニドリーノ" },
-  { id: 34, name: "ニドキング" },
-  { id: 35, name: "ピッピ" },
-  { id: 36, name: "ピクシー" },
-  { id: 37, name: "ロコン" },
-  { id: 38, name: "キュウコン" },
-  { id: 39, name: "プリン" },
-  { id: 40, name: "プクリン" },
-  { id: 41, name: "ズバット" },
-  { id: 42, name: "ゴルバット" },
-  { id: 43, name: "ナゾノクサ" },
-  { id: 44, name: "クサイハナ" },
-  { id: 45, name: "ラフレシア" },
-  { id: 46, name: "パラス" },
-  { id: 47, name: "パラセクト" },
-  { id: 48, name: "コンパン" },
-  { id: 49, name: "モルフォン" },
-  { id: 50, name: "ディグダ" },
-  { id: 51, name: "ダグトリオ" },
-  { id: 52, name: "ニャース" },
-  { id: 53, name: "ペルシアン" },
-  { id: 54, name: "コダック" },
-  { id: 55, name: "ゴルダック" },
-  { id: 56, name: "マンキー" },
-  { id: 57, name: "オコリザル" },
-  { id: 58, name: "ガーディ" },
-  { id: 59, name: "ウインディ" },
-  { id: 60, name: "ニョロモ" },
-  { id: 61, name: "ニョロゾ" },
-  { id: 62, name: "ニョロボン" },
-  { id: 63, name: "ケーシィ" },
-  { id: 64, name: "ユンゲラー" },
-  { id: 65, name: "フーディン" },
-  { id: 66, name: "ワンリキー" },
-  { id: 67, name: "ゴーリキー" },
-  { id: 68, name: "カイリキー" },
-  { id: 69, name: "マダツボミ" },
-  { id: 70, name: "ウツドン" },
-  { id: 71, name: "ウツボット" },
-  { id: 72, name: "メノクラゲ" },
-  { id: 73, name: "ドククラゲ" },
-  { id: 74, name: "イシツブテ" },
-  { id: 75, name: "ゴローン" },
-  { id: 76, name: "ゴローニャ" },
-  { id: 77, name: "ポニータ" },
-  { id: 78, name: "ギャロップ" },
-  { id: 79, name: "ヤドン" },
-  { id: 80, name: "ヤドラン" },
-  { id: 81, name: "コイル" },
-  { id: 82, name: "レアコイル" },
-  { id: 83, name: "カモネギ" },
-  { id: 84, name: "ドードー" },
-  { id: 85, name: "ドードリオ" },
-  { id: 86, name: "パウワウ" },
-  { id: 87, name: "ジュゴン" },
-  { id: 88, name: "ベトベター" },
-  { id: 89, name: "ベトベトン" },
-  { id: 90, name: "シェルダー" },
-  { id: 91, name: "パルシェン" },
-  { id: 92, name: "ゴース" },
-  { id: 93, name: "ゴースト" },
-  { id: 94, name: "ゲンガー" },
-  { id: 95, name: "イワーク" },
-  { id: 96, name: "スリープ" },
-  { id: 97, name: "スリーパー" },
-  { id: 98, name: "クラブ" },
-  { id: 99, name: "キングラー" },
-  { id: 100, name: "ビリリダマ" },
-  { id: 101, name: "マルマイン" },
-  { id: 102, name: "タマタマ" },
-  { id: 103, name: "ナッシー" },
-  { id: 104, name: "カラカラ" },
-  { id: 105, name: "ガラガラ" },
-  { id: 106, name: "サワムラー" },
-  { id: 107, name: "エビワラー" },
-  { id: 108, name: "ベロリンガ" },
-  { id: 109, name: "ドガース" },
-  { id: 110, name: "マタドガス" },
-  { id: 111, name: "サイホーン" },
-  { id: 112, name: "サイドン" },
-  { id: 113, name: "ラッキー" },
-  { id: 114, name: "モンジャラ" },
-  { id: 115, name: "ガルーラ" },
-  { id: 116, name: "タッツー" },
-  { id: 117, name: "シードラ" },
-  { id: 118, name: "トサキント" },
-  { id: 119, name: "アズマオウ" },
-  { id: 120, name: "ヒトデマン" },
-  { id: 121, name: "スターミー" },
-  { id: 122, name: "バリヤード" },
-  { id: 123, name: "ストライク" },
-  { id: 124, name: "ルージュラ" },
-  { id: 125, name: "エレブー" },
-  { id: 126, name: "ブーバー" },
-  { id: 127, name: "カイロス" },
-  { id: 128, name: "ケンタロス" },
-  { id: 129, name: "コイキング" },
-  { id: 130, name: "ギャラドス" },
-  { id: 131, name: "ラプラス" },
-  { id: 132, name: "メタモン" },
-  { id: 133, name: "イーブイ" },
-  { id: 134, name: "シャワーズ" },
-  { id: 135, name: "サンダース" },
-  { id: 136, name: "ブースター" },
-  { id: 137, name: "ポリゴン" },
-  { id: 138, name: "オムナイト" },
-  { id: 139, name: "オムスター" },
-  { id: 140, name: "カブト" },
-  { id: 141, name: "カブトプス" },
-  { id: 142, name: "プテラ" },
-  { id: 143, name: "カビゴン" },
-  { id: 144, name: "フリーザー" },
-  { id: 145, name: "サンダー" },
-  { id: 146, name: "ファイヤー" },
-  { id: 147, name: "ミニリュウ" },
-  { id: 148, name: "ハクリュー" },
-  { id: 149, name: "カイリュー" },
-  { id: 150, name: "ミュウツー" },
-  { id: 151, name: "ミュウ" },
-  { id: 152, name: "チコリータ" },
-  { id: 153, name: "ベイリーフ" },
-  { id: 154, name: "メガニウム" },
-  { id: 155, name: "ヒノアラシ" },
-  { id: 156, name: "マグマラシ" },
-  { id: 157, name: "バクフーン" },
-  { id: 158, name: "ワニノコ" },
-  { id: 159, name: "アリゲイツ" },
-  { id: 160, name: "オーダイル" },
-  { id: 161, name: "オタチ" },
-  { id: 162, name: "オオタチ" },
-  { id: 163, name: "ホーホー" },
-  { id: 164, name: "ヨルノズク" },
-  { id: 165, name: "レディバ" },
-  { id: 166, name: "レディアン" },
-  { id: 167, name: "イトマル" },
-  { id: 168, name: "アリアドス" },
-  { id: 169, name: "クロバット" },
-  { id: 170, name: "チョンチー" },
-  { id: 171, name: "ランターン" },
-  { id: 172, name: "ピチュー" },
-  { id: 173, name: "ピィ" },
-  { id: 174, name: "ププリン" },
-  { id: 175, name: "トゲピー" },
-  { id: 176, name: "トゲチック" },
-  { id: 177, name: "ネイティ" },
-  { id: 178, name: "ネイティオ" },
-  { id: 179, name: "メリープ" },
-  { id: 180, name: "モココ" },
-  { id: 181, name: "デンリュウ" },
-  { id: 182, name: "キレイハナ" },
-  { id: 183, name: "マリル" },
-  { id: 184, name: "マリルリ" },
-  { id: 185, name: "ウソッキー" },
-  { id: 186, name: "ニョロトノ" },
-  { id: 187, name: "ハネッコ" },
-  { id: 188, name: "ポポッコ" },
-  { id: 189, name: "ワタッコ" },
-  { id: 190, name: "エイパム" },
-  { id: 191, name: "ヒマナッツ" },
-  { id: 192, name: "キマワリ" },
-  { id: 193, name: "ヤンヤンマ" },
-  { id: 194, name: "ウパー" },
-  { id: 195, name: "ヌオー" },
-  { id: 196, name: "エーフィ" },
-  { id: 197, name: "ブラッキー" },
-  { id: 198, name: "ヤミカラス" },
-  { id: 199, name: "ヤドキング" },
-  { id: 200, name: "ムウマ" },
-  { id: 201, name: "アンノーン" },
-  { id: 202, name: "ソーナンス" },
-  { id: 203, name: "キリンリキ" },
-  { id: 204, name: "クヌギダマ" },
-  { id: 205, name: "フォレトス" },
-  { id: 206, name: "ノコッチ" },
-  { id: 207, name: "グライガー" },
-  { id: 208, name: "ハガネール" },
-  { id: 209, name: "ブルー" },
-  { id: 210, name: "グランブル" },
-  { id: 211, name: "ハリーセン" },
-  { id: 212, name: "ハッサム" },
-  { id: 213, name: "ツボツボ" },
-  { id: 214, name: "ヘラクロス" },
-  { id: 215, name: "ニューラ" },
-  { id: 216, name: "ヒメグマ" },
-  { id: 217, name: "リングマ" },
-  { id: 218, name: "マグマッグ" },
-  { id: 219, name: "マグカルゴ" },
-  { id: 220, name: "ウリムー" },
-  { id: 221, name: "イノムー" },
-  { id: 222, name: "サニーゴ" },
-  { id: 223, name: "テッポウオ" },
-  { id: 224, name: "オクタン" },
-  { id: 225, name: "デリバード" },
-  { id: 226, name: "マンタイン" },
-  { id: 227, name: "エアームド" },
-  { id: 228, name: "デルビル" },
-  { id: 229, name: "ヘルガー" },
-  { id: 230, name: "キングドラ" },
-  { id: 231, name: "ゴマゾウ" },
-  { id: 232, name: "ドンファン" },
-  { id: 233, name: "ポリゴン２" },
-  { id: 234, name: "オドシシ" },
-  { id: 235, name: "ドーブル" },
-  { id: 236, name: "バルキー" },
-  { id: 237, name: "カポエラー" },
-  { id: 238, name: "ムチュール" },
-  { id: 239, name: "エレキッド" },
-  { id: 240, name: "ブビィ" },
-  { id: 241, name: "ミルタンク" },
-  { id: 242, name: "ハピナス" },
-  { id: 243, name: "ライコウ" },
-  { id: 244, name: "エンテイ" },
-  { id: 245, name: "スイクン" },
-  { id: 246, name: "ヨーギラス" },
-  { id: 247, name: "サナギラス" },
-  { id: 248, name: "バンギラス" },
-  { id: 249, name: "ルギア" },
-  { id: 250, name: "ホウオウ" },
-  { id: 251, name: "セレビィ" },
-  { id: 252, name: "キモリ" },
-  { id: 253, name: "ジュプトル" },
-  { id: 254, name: "ジュカイン" },
-  { id: 255, name: "アチャモ" },
-  { id: 256, name: "ワカシャモ" },
-  { id: 257, name: "バシャーモ" },
-  { id: 258, name: "ミズゴロウ" },
-  { id: 259, name: "ヌマクロー" },
-  { id: 260, name: "ラグラージ" },
-  { id: 261, name: "ポチエナ" },
-  { id: 262, name: "グラエナ" },
-  { id: 263, name: "ジグザグマ" },
-  { id: 264, name: "マッスグマ" },
-  { id: 265, name: "ケムッソ" },
-  { id: 266, name: "カラサリス" },
-  { id: 267, name: "アゲハント" },
-  { id: 268, name: "マユルド" },
-  { id: 269, name: "ドクケイル" },
-  { id: 270, name: "ハスボー" },
-  { id: 271, name: "ハスブレロ" },
-  { id: 272, name: "ルンパッパ" },
-  { id: 273, name: "タネボー" },
-  { id: 274, name: "コノハナ" },
-  { id: 275, name: "ダーテング" },
-  { id: 276, name: "スバメ" },
-  { id: 277, name: "オオスバメ" },
-  { id: 278, name: "キャモメ" },
-  { id: 279, name: "ペリッパー" },
-  { id: 280, name: "ラルトス" },
-  { id: 281, name: "キルリア" },
-  { id: 282, name: "サーナイト" },
-  { id: 283, name: "アメタマ" },
-  { id: 284, name: "アメモース" },
-  { id: 285, name: "キノココ" },
-  { id: 286, name: "キノガッサ" },
-  { id: 287, name: "ナマケロ" },
-  { id: 288, name: "ヤルキモノ" },
-  { id: 289, name: "ケッキング" },
-  { id: 290, name: "ツチニン" },
-  { id: 291, name: "テッカニン" },
-  { id: 292, name: "ヌケニン" },
-  { id: 293, name: "ゴニョニョ" },
-  { id: 294, name: "ドゴーム" },
-  { id: 295, name: "バクオング" },
-  { id: 296, name: "マクノシタ" },
-  { id: 297, name: "ハリテヤマ" },
-  { id: 298, name: "ルリリ" },
-  { id: 299, name: "ノズパス" },
-  { id: 300, name: "エネコ" },
-  { id: 301, name: "エネコロロ" },
-  { id: 302, name: "ヤミラミ" },
-  { id: 303, name: "クチート" },
-  { id: 304, name: "ココドラ" },
-  { id: 305, name: "コドラ" },
-  { id: 306, name: "ボスゴドラ" },
-  { id: 307, name: "アサナン" },
-  { id: 308, name: "チャーレム" },
-  { id: 309, name: "ラクライ" },
-  { id: 310, name: "ライボルト" },
-  { id: 311, name: "プラスル" },
-  { id: 312, name: "マイナン" },
-  { id: 313, name: "バルビート" },
-  { id: 314, name: "イルミーゼ" },
-  { id: 315, name: "ロゼリア" },
-  { id: 316, name: "ゴクリン" },
-  { id: 317, name: "マルノーム" },
-  { id: 318, name: "キバニア" },
-  { id: 319, name: "サメハダー" },
-  { id: 320, name: "ホエルコ" },
-  { id: 321, name: "ホエルオー" },
-  { id: 322, name: "ドンメル" },
-  { id: 323, name: "バクーダ" },
-  { id: 324, name: "コータス" },
-  { id: 325, name: "バネブー" },
-  { id: 326, name: "ブーピッグ" },
-  { id: 327, name: "パッチール" },
-  { id: 328, name: "ナックラー" },
-  { id: 329, name: "ビブラーバ" },
-  { id: 330, name: "フライゴン" },
-  { id: 331, name: "サボネア" },
-  { id: 332, name: "ノクタス" },
-  { id: 333, name: "チルット" },
-  { id: 334, name: "チルタリス" },
-  { id: 335, name: "ザングース" },
-  { id: 336, name: "ハブネーク" },
-  { id: 337, name: "ルナトーン" },
-  { id: 338, name: "ソルロック" },
-  { id: 339, name: "ドジョッチ" },
-  { id: 340, name: "ナマズン" },
-  { id: 341, name: "ヘイガニ" },
-  { id: 342, name: "シザリガー" },
-  { id: 343, name: "ヤジロン" },
-  { id: 344, name: "ネンドール" },
-  { id: 345, name: "リリーラ" },
-  { id: 346, name: "ユレイドル" },
-  { id: 347, name: "アノプス" },
-  { id: 348, name: "アーマルド" },
-  { id: 349, name: "ヒンバス" },
-  { id: 350, name: "ミロカロス" },
-  { id: 351, name: "ポワルン" },
-  { id: 352, name: "カクレオン" },
-  { id: 353, name: "カゲボウズ" },
-  { id: 354, name: "ジュペッタ" },
-  { id: 355, name: "ヨマワル" },
-  { id: 356, name: "サマヨール" },
-  { id: 357, name: "トロピウス" },
-  { id: 358, name: "チリーン" },
-  { id: 359, name: "アブソル" },
-  { id: 360, name: "ソーナノ" },
-  { id: 361, name: "ユキワラシ" },
-  { id: 362, name: "オニゴーリ" },
-  { id: 363, name: "タマザラシ" },
-  { id: 364, name: "トドグラー" },
-  { id: 365, name: "トドゼルガ" },
-  { id: 366, name: "パールル" },
-  { id: 367, name: "ハンテール" },
-  { id: 368, name: "サクラビス" },
-  { id: 369, name: "ジーランス" },
-  { id: 370, name: "ラブカス" },
-  { id: 371, name: "タツベイ" },
-  { id: 372, name: "コモルー" },
-  { id: 373, name: "ボーマンダ" },
-  { id: 374, name: "ダンバル" },
-  { id: 375, name: "メタング" },
-  { id: 376, name: "メタグロス" },
-  { id: 377, name: "レジロック" },
-  { id: 378, name: "レジアイス" },
-  { id: 379, name: "レジスチル" },
-  { id: 380, name: "ラティアス" },
-  { id: 381, name: "ラティオス" },
-  { id: 382, name: "カイオーガ" },
-  { id: 383, name: "グラードン" },
-  { id: 384, name: "レックウザ" },
-  { id: 385, name: "ジラーチ" },
-  { id: 386, name: "デオキシス" },
-  { id: 387, name: "ナエトル" },
-  { id: 388, name: "ハヤシガメ" },
-  { id: 389, name: "ドダイトス" },
-  { id: 390, name: "ヒコザル" },
-  { id: 391, name: "モウカザル" },
-  { id: 392, name: "ゴウカザル" },
-  { id: 393, name: "ポッチャマ" },
-  { id: 394, name: "ポッタイシ" },
-  { id: 395, name: "エンペルト" },
-  { id: 396, name: "ムックル" },
-  { id: 397, name: "ムクバード" },
-  { id: 398, name: "ムクホーク" },
-  { id: 399, name: "ビッパ" },
-  { id: 400, name: "ビーダル" },
-  { id: 401, name: "コロボーシ" },
-  { id: 402, name: "コロトック" },
-  { id: 403, name: "コリンク" },
-  { id: 404, name: "ルクシオ" },
-  { id: 405, name: "レントラー" },
-  { id: 406, name: "スボミー" },
-  { id: 407, name: "ロズレイド" },
-  { id: 408, name: "ズガイドス" },
-  { id: 409, name: "ラムパルド" },
-  { id: 410, name: "タテトプス" },
-  { id: 411, name: "トリデプス" },
-  { id: 412, name: "ミノムッチ" },
-  { id: 413, name: "ミノマダム" },
-  { id: 414, name: "ガーメイル" },
-  { id: 415, name: "ミツハニー" },
-  { id: 416, name: "ビークイン" },
-  { id: 417, name: "パチリス" },
-  { id: 418, name: "ブイゼル" },
-  { id: 419, name: "フローゼル" },
-  { id: 420, name: "チェリンボ" },
-  { id: 421, name: "チェリム" },
-  { id: 422, name: "カラナクシ" },
-  { id: 423, name: "トリトドン" },
-  { id: 424, name: "エテボース" },
-  { id: 425, name: "フワンテ" },
-  { id: 426, name: "フワライド" },
-  { id: 427, name: "ミミロル" },
-  { id: 428, name: "ミミロップ" },
-  { id: 429, name: "ムウマージ" },
-  { id: 430, name: "ドンカラス" },
-  { id: 431, name: "ニャルマー" },
-  { id: 432, name: "ブニャット" },
-  { id: 433, name: "リーシャン" },
-  { id: 434, name: "スカンプー" },
-  { id: 435, name: "スカタンク" },
-  { id: 436, name: "ドーミラー" },
-  { id: 437, name: "ドータクン" },
-  { id: 438, name: "ウソハチ" },
-  { id: 439, name: "マネネ" },
-  { id: 440, name: "ピンプク" },
-  { id: 441, name: "ペラップ" },
-  { id: 442, name: "ミカルゲ" },
-  { id: 443, name: "フカマル" },
-  { id: 444, name: "ガバイト" },
-  { id: 445, name: "ガブリアス" },
-  { id: 446, name: "ゴンベ" },
-  { id: 447, name: "リオル" },
-  { id: 448, name: "ルカリオ" },
-  { id: 449, name: "ヒポポタス" },
-  { id: 450, name: "カバルドン" },
-  { id: 451, name: "スコルピ" },
-  { id: 452, name: "ドラピオン" },
-  { id: 453, name: "グレッグル" },
-  { id: 454, name: "ドクロッグ" },
-  { id: 455, name: "マスキッパ" },
-  { id: 456, name: "ケイコウオ" },
-  { id: 457, name: "ネオラント" },
-  { id: 458, name: "タマンタ" },
-  { id: 459, name: "ユキカブリ" },
-  { id: 460, name: "ユキノオー" },
-  { id: 461, name: "マニューラ" },
-  { id: 462, name: "ジバコイル" },
-  { id: 463, name: "ベロベルト" },
-  { id: 464, name: "ドサイドン" },
-  { id: 465, name: "モジャンボ" },
-  { id: 466, name: "エレキブル" },
-  { id: 467, name: "ブーバーン" },
-  { id: 468, name: "トゲキッス" },
-  { id: 469, name: "メガヤンマ" },
-  { id: 470, name: "リーフィア" },
-  { id: 471, name: "グレイシア" },
-  { id: 472, name: "グライオン" },
-  { id: 473, name: "マンムー" },
-  { id: 474, name: "ポリゴンＺ" },
-  { id: 475, name: "エルレイド" },
-  { id: 476, name: "ダイノーズ" },
-  { id: 477, name: "ヨノワール" },
-  { id: 478, name: "ユキメノコ" },
-  { id: 479, name: "ロトム" },
-  { id: 480, name: "ユクシー" },
-  { id: 481, name: "エムリット" },
-  { id: 482, name: "アグノム" },
-  { id: 483, name: "ディアルガ" },
-  { id: 484, name: "パルキア" },
-  { id: 485, name: "ヒードラン" },
-  { id: 486, name: "レジギガス" },
-  { id: 487, name: "ギラティナ" },
-  { id: 488, name: "クレセリア" },
-  { id: 489, name: "フィオネ" },
-  { id: 490, name: "マナフィ" },
-  { id: 491, name: "ダークライ" },
-  { id: 492, name: "シェイミ" },
-  { id: 493, name: "アルセウス" },
-  { id: 494, name: "ビクティニ" },
-  { id: 495, name: "ツタージャ" },
-  { id: 496, name: "ジャノビー" },
-  { id: 497, name: "ジャローダ" },
-  { id: 498, name: "ポカブ" },
-  { id: 499, name: "チャオブー" },
-  { id: 500, name: "エンブオー" },
-  { id: 501, name: "ミジュマル" },
-  { id: 502, name: "フタチマル" },
-  { id: 503, name: "ダイケンキ" },
-  { id: 504, name: "ミネズミ" },
-  { id: 505, name: "ミルホッグ" },
-  { id: 506, name: "ヨーテリー" },
-  { id: 507, name: "ハーデリア" },
-  { id: 508, name: "ムーランド" },
-  { id: 509, name: "チョロネコ" },
-  { id: 510, name: "レパルダス" },
-  { id: 511, name: "ヤナップ" },
-  { id: 512, name: "ヤナッキー" },
-  { id: 513, name: "バオップ" },
-  { id: 514, name: "バオッキー" },
-  { id: 515, name: "ヒヤップ" },
-  { id: 516, name: "ヒヤッキー" },
-  { id: 517, name: "ムンナ" },
-  { id: 518, name: "ムシャーナ" },
-  { id: 519, name: "マメパト" },
-  { id: 520, name: "ハトーボー" },
-  { id: 521, name: "ケンホロウ" },
-  { id: 522, name: "シママ" },
-  { id: 523, name: "ゼブライカ" },
-  { id: 524, name: "ダンゴロ" },
-  { id: 525, name: "ガントル" },
-  { id: 526, name: "ギガイアス" },
-  { id: 527, name: "コロモリ" },
-  { id: 528, name: "ココロモリ" },
-  { id: 529, name: "モグリュー" },
-  { id: 530, name: "ドリュウズ" },
-  { id: 531, name: "タブンネ" },
-  { id: 532, name: "ドッコラー" },
-  { id: 533, name: "ドテッコツ" },
-  { id: 534, name: "ローブシン" },
-  { id: 535, name: "オタマロ" },
-  { id: 536, name: "ガマガル" },
-  { id: 537, name: "ガマゲロゲ" },
-  { id: 538, name: "ナゲキ" },
-  { id: 539, name: "ダゲキ" },
-  { id: 540, name: "クルミル" },
-  { id: 541, name: "クルマユ" },
-  { id: 542, name: "ハハコモリ" },
-  { id: 543, name: "フシデ" },
-  { id: 544, name: "ホイーガ" },
-  { id: 545, name: "ペンドラー" },
-  { id: 546, name: "モンメン" },
-  { id: 547, name: "エルフーン" },
-  { id: 548, name: "チュリネ" },
-  { id: 549, name: "ドレディア" },
-  { id: 550, name: "バスラオ" },
-  { id: 551, name: "メグロコ" },
-  { id: 552, name: "ワルビル" },
-  { id: 553, name: "ワルビアル" },
-  { id: 554, name: "ダルマッカ" },
-  { id: 555, name: "ヒヒダルマ" },
-  { id: 556, name: "マラカッチ" },
-  { id: 557, name: "イシズマイ" },
-  { id: 558, name: "イワパレス" },
-  { id: 559, name: "ズルッグ" },
-  { id: 560, name: "ズルズキン" },
-  { id: 561, name: "シンボラー" },
-  { id: 562, name: "デスマス" },
-  { id: 563, name: "デスカーン" },
-  { id: 564, name: "プロトーガ" },
-  { id: 565, name: "アバゴーラ" },
-  { id: 566, name: "アーケン" },
-  { id: 567, name: "アーケオス" },
-  { id: 568, name: "ヤブクロン" },
-  { id: 569, name: "ダストダス" },
-  { id: 570, name: "ゾロア" },
-  { id: 571, name: "ゾロアーク" },
-  { id: 572, name: "チラーミィ" },
-  { id: 573, name: "チラチーノ" },
-  { id: 574, name: "ゴチム" },
-  { id: 575, name: "ゴチミル" },
-  { id: 576, name: "ゴチルゼル" },
-  { id: 577, name: "ユニラン" },
-  { id: 578, name: "ダブラン" },
-  { id: 579, name: "ランクルス" },
-  { id: 580, name: "コアルヒー" },
-  { id: 581, name: "スワンナ" },
-  { id: 582, name: "バニプッチ" },
-  { id: 583, name: "バニリッチ" },
-  { id: 584, name: "バイバニラ" },
-  { id: 585, name: "シキジカ" },
-  { id: 586, name: "メブキジカ" },
-  { id: 587, name: "エモンガ" },
-  { id: 588, name: "カブルモ" },
-  { id: 589, name: "シュバルゴ" },
-  { id: 590, name: "タマゲタケ" },
-  { id: 591, name: "モロバレル" },
-  { id: 592, name: "プルリル" },
-  { id: 593, name: "ブルンゲル" },
-  { id: 594, name: "ママンボウ" },
-  { id: 595, name: "バチュル" },
-  { id: 596, name: "デンチュラ" },
-  { id: 597, name: "テッシード" },
-  { id: 598, name: "ナットレイ" },
-  { id: 599, name: "ギアル" },
-  { id: 600, name: "ギギアル" },
-  { id: 601, name: "ギギギアル" },
-  { id: 602, name: "シビシラス" },
-  { id: 603, name: "シビビール" },
-  { id: 604, name: "シビルドン" },
-  { id: 605, name: "リグレー" },
-  { id: 606, name: "オーベム" },
-  { id: 607, name: "ヒトモシ" },
-  { id: 608, name: "ランプラー" },
-  { id: 609, name: "シャンデラ" },
-  { id: 610, name: "キバゴ" },
-  { id: 611, name: "オノンド" },
-  { id: 612, name: "オノノクス" },
-  { id: 613, name: "クマシュン" },
-  { id: 614, name: "ツンベアー" },
-  { id: 615, name: "フリージオ" },
-  { id: 616, name: "チョボマキ" },
-  { id: 617, name: "アギルダー" },
-  { id: 618, name: "マッギョ" },
-  { id: 619, name: "コジョフー" },
-  { id: 620, name: "コジョンド" },
-  { id: 621, name: "クリムガン" },
-  { id: 622, name: "ゴビット" },
-  { id: 623, name: "ゴルーグ" },
-  { id: 624, name: "コマタナ" },
-  { id: 625, name: "キリキザン" },
-  { id: 626, name: "バッフロン" },
-  { id: 627, name: "ワシボン" },
-  { id: 628, name: "ウォーグル" },
-  { id: 629, name: "バルチャイ" },
-  { id: 630, name: "バルジーナ" },
-  { id: 631, name: "クイタラン" },
-  { id: 632, name: "アイアント" },
-  { id: 633, name: "モノズ" },
-  { id: 634, name: "ジヘッド" },
-  { id: 635, name: "サザンドラ" },
-  { id: 636, name: "メラルバ" },
-  { id: 637, name: "ウルガモス" },
-  { id: 638, name: "コバルオン" },
-  { id: 639, name: "テラキオン" },
-  { id: 640, name: "ビリジオン" },
-  { id: 641, name: "トルネロス" },
-  { id: 642, name: "ボルトロス" },
-  { id: 643, name: "レシラム" },
-  { id: 644, name: "ゼクロム" },
-  { id: 645, name: "ランドロス" },
-  { id: 646, name: "キュレム" },
-  { id: 647, name: "ケルディオ" },
-  { id: 648, name: "メロエッタ" },
-  { id: 649, name: "ゲノセクト" },
-  { id: 650, name: "ハリマロン" },
-  { id: 651, name: "ハリボーグ" },
-  { id: 652, name: "ブリガロン" },
-  { id: 653, name: "フォッコ" },
-  { id: 654, name: "テールナー" },
-  { id: 655, name: "マフォクシー" },
-  { id: 656, name: "ケロマツ" },
-  { id: 657, name: "ゲコガシラ" },
-  { id: 658, name: "ゲッコウガ" },
-  { id: 659, name: "ホルビー" },
-  { id: 660, name: "ホルード" },
-  { id: 661, name: "ヤヤコマ" },
-  { id: 662, name: "ヒノヤコマ" },
-  { id: 663, name: "ファイアロー" },
-  { id: 664, name: "コフキムシ" },
-  { id: 665, name: "コフーライ" },
-  { id: 666, name: "ビビヨン" },
-  { id: 667, name: "シシコ" },
-  { id: 668, name: "カエンジシ" },
-  { id: 669, name: "フラベベ" },
-  { id: 670, name: "フラエッテ" },
-  { id: 671, name: "フラージェス" },
-  { id: 672, name: "メェークル" },
-  { id: 673, name: "ゴーゴート" },
-  { id: 674, name: "ヤンチャム" },
-  { id: 675, name: "ゴロンダ" },
-  { id: 676, name: "トリミアン" },
-  { id: 677, name: "ニャスパー" },
-  { id: 678, name: "ニャオニクス" },
-  { id: 679, name: "ヒトツキ" },
-  { id: 680, name: "ニダンギル" },
-  { id: 681, name: "ギルガルド" },
-  { id: 682, name: "シュシュプ" },
-  { id: 683, name: "フレフワン" },
-  { id: 684, name: "ペロッパフ" },
-  { id: 685, name: "ペロリーム" },
-  { id: 686, name: "マーイーカ" },
-  { id: 687, name: "カラマネロ" },
-  { id: 688, name: "カメテテ" },
-  { id: 689, name: "ガメノデス" },
-  { id: 690, name: "クズモー" },
-  { id: 691, name: "ドラミドロ" },
-  { id: 692, name: "ウデッポウ" },
-  { id: 693, name: "ブロスター" },
-  { id: 694, name: "エリキテル" },
-  { id: 695, name: "エレザード" },
-  { id: 696, name: "チゴラス" },
-  { id: 697, name: "ガチゴラス" },
-  { id: 698, name: "アマルス" },
-  { id: 699, name: "アマルルガ" },
-  { id: 700, name: "ニンフィア" },
-  { id: 701, name: "ルチャブル" },
-  { id: 702, name: "デデンネ" },
-  { id: 703, name: "メレシー" },
-  { id: 704, name: "ヌメラ" },
-  { id: 705, name: "ヌメイル" },
-  { id: 706, name: "ヌメルゴン" },
-  { id: 707, name: "クレッフィ" },
-  { id: 708, name: "ボクレー" },
-  { id: 709, name: "オーロット" },
-  { id: 710, name: "バケッチャ" },
-  { id: 711, name: "パンプジン" },
-  { id: 712, name: "カチコール" },
-  { id: 713, name: "クレベース" },
-  { id: 714, name: "オンバット" },
-  { id: 715, name: "オンバーン" },
-  { id: 716, name: "ゼルネアス" },
-  { id: 717, name: "イベルタル" },
-  { id: 718, name: "ジガルデ" },
-  { id: 719, name: "ディアンシー" },
-  { id: 720, name: "フーパ" },
-  { id: 721, name: "ボルケニオン" },
-  { id: 722, name: "モクロー" },
-  { id: 723, name: "フクスロー" },
-  { id: 724, name: "ジュナイパー" },
-  { id: 725, name: "ニャビー" },
-  { id: 726, name: "ニャヒート" },
-  { id: 727, name: "ガオガエン" },
-  { id: 728, name: "アシマリ" },
-  { id: 729, name: "オシャマリ" },
-  { id: 730, name: "アシレーヌ" },
-  { id: 731, name: "ツツケラ" },
-  { id: 732, name: "ケララッパ" },
-  { id: 733, name: "ドデカバシ" },
-  { id: 734, name: "ヤングース" },
-  { id: 735, name: "デカグース" },
-  { id: 736, name: "アゴジムシ" },
-  { id: 737, name: "デンヂムシ" },
-  { id: 738, name: "クワガノン" },
-  { id: 739, name: "マケンカニ" },
-  { id: 740, name: "ケケンカニ" },
-  { id: 741, name: "オドリドリ" },
-  { id: 742, name: "アブリー" },
-  { id: 743, name: "アブリボン" },
-  { id: 744, name: "イワンコ" },
-  { id: 745, name: "ルガルガン" },
-  { id: 746, name: "ヨワシ" },
-  { id: 747, name: "ヒドイデ" },
-  { id: 748, name: "ドヒドイデ" },
-  { id: 749, name: "ドロバンコ" },
-  { id: 750, name: "バンバドロ" },
-  { id: 751, name: "シズクモ" },
-  { id: 752, name: "オニシズクモ" },
-  { id: 753, name: "カリキリ" },
-  { id: 754, name: "ラランテス" },
-  { id: 755, name: "ネマシュ" },
-  { id: 756, name: "マシェード" },
-  { id: 757, name: "ヤトウモリ" },
-  { id: 758, name: "エンニュート" },
-  { id: 759, name: "ヌイコグマ" },
-  { id: 760, name: "キテルグマ" },
-  { id: 761, name: "アマカジ" },
-  { id: 762, name: "アママイコ" },
-  { id: 763, name: "アマージョ" },
-  { id: 764, name: "キュワワー" },
-  { id: 765, name: "ヤレユータン" },
-  { id: 766, name: "ナゲツケサル" },
-  { id: 767, name: "コソクムシ" },
-  { id: 768, name: "グソクムシャ" },
-  { id: 769, name: "スナバァ" },
-  { id: 770, name: "シロデスナ" },
-  { id: 771, name: "ナマコブシ" },
-  { id: 772, name: "タイプ：ヌル" },
-  { id: 773, name: "シルヴァディ" },
-  { id: 774, name: "メテノ" },
-  { id: 775, name: "ネッコアラ" },
-  { id: 776, name: "バクガメス" },
-  { id: 777, name: "トゲデマル" },
-  { id: 778, name: "ミミッキュ" },
-  { id: 779, name: "ハギギシリ" },
-  { id: 780, name: "ジジーロン" },
-  { id: 781, name: "ダダリン" },
-  { id: 782, name: "ジャラコ" },
-  { id: 783, name: "ジャランゴ" },
-  { id: 784, name: "ジャラランガ" },
-  { id: 785, name: "カプ・コケコ" },
-  { id: 786, name: "カプ・テテフ" },
-  { id: 787, name: "カプ・ブルル" },
-  { id: 788, name: "カプ・レヒレ" },
-  { id: 789, name: "コスモッグ" },
-  { id: 790, name: "コスモウム" },
-  { id: 791, name: "ソルガレオ" },
-  { id: 792, name: "ルナアーラ" },
-  { id: 793, name: "ウツロイド" },
-  { id: 794, name: "マッシブーン" },
-  { id: 795, name: "フェローチェ" },
-  { id: 796, name: "デンジュモク" },
-  { id: 797, name: "テッカグヤ" },
-  { id: 798, name: "カミツルギ" },
-  { id: 799, name: "アクジキング" },
-  { id: 800, name: "ネクロズマ" },
-  { id: 801, name: "マギアナ" },
-  { id: 802, name: "マーシャドー" },
-  { id: 803, name: "ベベノム" },
-  { id: 804, name: "アーゴヨン" },
-  { id: 805, name: "ツンデツンデ" },
-  { id: 806, name: "ズガドーン" },
-  { id: 807, name: "ゼラオラ" },
-  { id: 808, name: "メルタン" },
-  { id: 809, name: "メルメタル" },
-  { id: 810, name: "サルノリ" },
-  { id: 811, name: "バチンキー" },
-  { id: 812, name: "ゴリランダー" },
-  { id: 813, name: "ヒバニー" },
-  { id: 814, name: "ラビフット" },
-  { id: 815, name: "エースバーン" },
-  { id: 816, name: "メッソン" },
-  { id: 817, name: "ジメレオン" },
-  { id: 818, name: "インテレオン" },
-  { id: 819, name: "ホシガリス" },
-  { id: 820, name: "ヨクバリス" },
-  { id: 821, name: "ココガラ" },
-  { id: 822, name: "アオガラス" },
-  { id: 823, name: "アーマーガア" },
-  { id: 824, name: "サッチムシ" },
-  { id: 825, name: "レドームシ" },
-  { id: 826, name: "イオルブ" },
-  { id: 827, name: "クスネ" },
-  { id: 828, name: "フォクスライ" },
-  { id: 829, name: "ヒメンカ" },
-  { id: 830, name: "ワタシラガ" },
-  { id: 831, name: "ウールー" },
-  { id: 832, name: "バイウールー" },
-  { id: 833, name: "カムカメ" },
-  { id: 834, name: "カジリガメ" },
-  { id: 835, name: "ワンパチ" },
-  { id: 836, name: "パルスワン" },
-  { id: 837, name: "タンドン" },
-  { id: 838, name: "トロッゴン" },
-  { id: 839, name: "セキタンザン" },
-  { id: 840, name: "カジッチュ" },
-  { id: 841, name: "アップリュー" },
-  { id: 842, name: "タルップル" },
-  { id: 843, name: "スナヘビ" },
-  { id: 844, name: "サダイジャ" },
-  { id: 845, name: "ウッウ" },
-  { id: 846, name: "サシカマス" },
-  { id: 847, name: "カマスジョー" },
-  { id: 848, name: "エレズン" },
-  { id: 849, name: "ストリンダー" },
-  { id: 850, name: "ヤクデ" },
-  { id: 851, name: "マルヤクデ" },
-  { id: 852, name: "タタッコ" },
-  { id: 853, name: "オトスパス" },
-  { id: 854, name: "ヤバチャ" },
-  { id: 855, name: "ポットデス" },
-  { id: 856, name: "ミブリム" },
-  { id: 857, name: "テブリム" },
-  { id: 858, name: "ブリムオン" },
-  { id: 859, name: "ベロバー" },
-  { id: 860, name: "ギモー" },
-  { id: 861, name: "オーロンゲ" },
-  { id: 862, name: "タチフサグマ" },
-  { id: 863, name: "ニャイキング" },
-  { id: 864, name: "サニゴーン" },
-  { id: 865, name: "ネギガナイト" },
-  { id: 866, name: "バリコオル" },
-  { id: 867, name: "デスバーン" },
-  { id: 868, name: "マホミル" },
-  { id: 869, name: "マホイップ" },
-  { id: 870, name: "タイレーツ" },
-  { id: 871, name: "バチンウニ" },
-  { id: 872, name: "ユキハミ" },
-  { id: 873, name: "モスノウ" },
-  { id: 874, name: "イシヘンジン" },
-  { id: 875, name: "コオリッポ" },
-  { id: 876, name: "イエッサン" },
-  { id: 877, name: "モルペコ" },
-  { id: 878, name: "ゾウドウ" },
-  { id: 879, name: "ダイオウドウ" },
-  { id: 880, name: "パッチラゴン" },
-  { id: 881, name: "パッチルドン" },
-  { id: 882, name: "ウオノラゴン" },
-  { id: 883, name: "ウオチルドン" },
-  { id: 884, name: "ジュラルドン" },
-  { id: 885, name: "ドラメシヤ" },
-  { id: 886, name: "ドロンチ" },
-  { id: 887, name: "ドラパルト" },
-  { id: 888, name: "ザシアン" },
-  { id: 889, name: "ザマゼンタ" },
-  { id: 890, name: "ムゲンダイナ" },
-  { id: 891, name: "ダクマ" },
-  { id: 892, name: "ウーラオス" },
-  { id: 893, name: "ザルード" },
-  { id: 894, name: "レジエレキ" },
-  { id: 895, name: "レジドラゴ" },
-  { id: 896, name: "ブリザポス" },
-  { id: 897, name: "レイスポス" },
-  { id: 898, name: "バドレックス" },
-  { id: 899, name: "アヤシシ" },
-  { id: 900, name: "バサギリ" },
-  { id: 901, name: "ガチグマ" },
-  { id: 902, name: "イダイトウ" },
-  { id: 903, name: "オオニューラ" },
-  { id: 904, name: "ハリーマン" },
-  { id: 905, name: "ラブトロス" },
-  { id: 906, name: "ニャオハ" },
-  { id: 907, name: "ニャローテ" },
-  { id: 908, name: "マスカーニャ" },
-  { id: 909, name: "ホゲータ" },
-  { id: 910, name: "アチゲータ" },
-  { id: 911, name: "ラウドボーン" },
-  { id: 912, name: "クワッス" },
-  { id: 913, name: "ウェルカモ" },
-  { id: 914, name: "ウェーニバル" },
-  { id: 915, name: "グルトン" },
-  { id: 916, name: "パフュートン" },
-  { id: 917, name: "タマンチュラ" },
-  { id: 918, name: "ワナイダー" },
-  { id: 919, name: "マメバッタ" },
-  { id: 920, name: "エクスレッグ" },
-  { id: 921, name: "パモ" },
-  { id: 922, name: "パモット" },
-  { id: 923, name: "パーモット" },
-  { id: 924, name: "ワッカネズミ" },
-  { id: 925, name: "イッカネズミ" },
-  { id: 926, name: "パピモッチ" },
-  { id: 927, name: "バウッツェル" },
-  { id: 928, name: "ミニーブ" },
-  { id: 929, name: "オリーニョ" },
-  { id: 930, name: "オリーヴァ" },
-  { id: 931, name: "イキリンコ" },
-  { id: 932, name: "コジオ" },
-  { id: 933, name: "ジオヅム" },
-  { id: 934, name: "キョジオーン" },
-  { id: 935, name: "カルボウ" },
-  { id: 936, name: "グレンアルマ" },
-  { id: 937, name: "ソウブレイズ" },
-  { id: 938, name: "ズピカ" },
-  { id: 939, name: "ハラバリー" },
-  { id: 940, name: "カイデン" },
-  { id: 941, name: "タイカイデン" },
-  { id: 942, name: "オラチフ" },
-  { id: 943, name: "マフィティフ" },
-  { id: 944, name: "シルシュルー" },
-  { id: 945, name: "タギングル" },
-  { id: 946, name: "アノクサ" },
-  { id: 947, name: "アノホラグサ" },
-  { id: 948, name: "ノノクラゲ" },
-  { id: 949, name: "リククラゲ" },
-  { id: 950, name: "ガケガニ" },
-  { id: 951, name: "カプサイジ" },
-  { id: 952, name: "スコヴィラン" },
-  { id: 953, name: "シガロコ" },
-  { id: 954, name: "ベラカス" },
-  { id: 955, name: "ヒラヒナ" },
-  { id: 956, name: "クエスパトラ" },
-  { id: 957, name: "カヌチャン" },
-  { id: 958, name: "ナカヌチャン" },
-  { id: 959, name: "デカヌチャン" },
-  { id: 960, name: "ウミディグダ" },
-  { id: 961, name: "ウミトリオ" },
-  { id: 962, name: "オトシドリ" },
-  { id: 963, name: "ナミイルカ" },
-  { id: 964, name: "イルカマン" },
-  { id: 965, name: "ブロロン" },
-  { id: 966, name: "ブロロローム" },
-  { id: 967, name: "モトトカゲ" },
-  { id: 968, name: "ミミズズ" },
-  { id: 969, name: "キラーメ" },
-  { id: 970, name: "キラフロル" },
-  { id: 971, name: "ボチ" },
-  { id: 972, name: "ハカドッグ" },
-  { id: 973, name: "カラミンゴ" },
-  { id: 974, name: "アルクジラ" },
-  { id: 975, name: "ハルクジラ" },
-  { id: 976, name: "ミガルーサ" },
-  { id: 977, name: "ヘイラッシャ" },
-  { id: 978, name: "シャリタツ" },
-  { id: 979, name: "コノヨザル" },
-  { id: 980, name: "ドオー" },
-  { id: 981, name: "リキキリン" },
-  { id: 982, name: "ノココッチ" },
-  { id: 983, name: "ドドゲザン" },
-  { id: 984, name: "イダイナキバ" },
-  { id: 985, name: "サケブシッポ" },
-  { id: 986, name: "アラブルタケ" },
-  { id: 987, name: "ハバタクカミ" },
-  { id: 988, name: "チヲハウハネ" },
-  { id: 989, name: "スナノケガワ" },
-  { id: 990, name: "テツノワダチ" },
-  { id: 991, name: "テツノツツミ" },
-  { id: 992, name: "テツノカイナ" },
-  { id: 993, name: "テツノコウベ" },
-  { id: 994, name: "テツノドクガ" },
-  { id: 995, name: "テツノイバラ" },
-  { id: 996, name: "セビエ" },
-  { id: 997, name: "セゴール" },
-  { id: 998, name: "セグレイブ" },
-  { id: 999, name: "コレクレー" },
-  { id: 1000, name: "サーフゴー" },
+export const LINK_LIST = [
+  {
+    text: "Pokémon List",
+    href: "/pokemon",
+  },
+  {
+    text: "Search",
+    href: "/pokemon/search",
+  },
 ];
